PPI-1357 Move domains config functions to clientConfiguration.service
