+++ src/services/cross-domain/crossDomain.service.ts
@@ -71,23 +71,3 @@ export function customCrossDomainLinkVisitorIdGetter(getter: VisitorIdGetter) {
     },
   ])
 }
-
-export function setDomains(domains: string[]) {
-  push([CROSS_DOMAIN_TRACK_EVENT.SET_DOMAINS, domains])
-}
-
-export function getDomains(): Promise<string[]> {
-  return new Promise<string[]>((resolve, reject) => {
-    try {
-      push([
-        function (this: Tracker) {
-          resolve(this.getDomains())
-        },
-      ])
-    } catch (e) {
-      if (e instanceof ReferenceError) {
-        reject(e)
-      }
-    }
-  })
-}
